refactor(api): neutralize slack connect and data service naming
Rename the two Slack API service modules and their branded exported
identifiers to domain-neutral names, updating every caller atomically
with no compatibility alias and no old-path re-export.

File renames:
- services/zero-slack-connect.service.ts -> services/slack-connect.service.ts
- services/zero-slack-data.service.ts -> services/slack-data.service.ts

Identifier renames:
- zeroSlackConnectStatus -> slackConnectStatus
- zeroSlackOrgStatus -> slackOrgStatus
- zeroSlackOrgInstallation -> slackOrgInstallation
- zeroSlackChannels -> slackChannels

The two modules share callers, so both move together. connectSlackWorkspace$,
publishSlackAdminSignal$, notifySlackConnect$ and SLACK_BOT_SCOPES were
already neutral and are unchanged; every SLACK_BOT_SCOPES value stays
byte-identical.

The agent-compose consolidation preflight repository manifest pins scanned
repository paths and is asserted against a live repository scan, so its five
entries for the two renamed services move with them and are re-sorted in
collector order. The entry contents are otherwise unchanged.

The platform computed slackConnectStatus$ is a different package and a
different symbol, and is untouched. The api-contracts package is untouched.

No Slack scope string, event type, API method name, wire payload key, HTTP
path literal, contract shape, database identifier or runtime behavior changes.

Closes #27916
