+++ turbo/apps/api/src/signals/routes/integrations-slack.ts
@@ -28,10 +28,10 @@ import { authRoute } from "../auth/auth-route";
 import { queryOf } from "../context/request";
 import { publicBrand$ } from "../context/hono";
 import {
-  zeroSlackOrgInstallation,
-  zeroSlackOrgStatus,
-} from "../services/zero-slack-data.service";
-import { publishSlackAdminSignal$ } from "../services/zero-slack-connect.service";
+  slackOrgInstallation,
+  slackOrgStatus,
+} from "../services/slack-data.service";
+import { publishSlackAdminSignal$ } from "../services/slack-connect.service";
 import { getFileInfo, isSlackApiClientError } from "../../lib/slack-client";
 import {
   fetchSlackFile,
@@ -178,7 +178,7 @@ const getSlackStatusInner$ = computed(async (get) => {
   const auth = get(organizationAuthContext$);
   const publicBrand = get(publicBrand$);
   const status = await get(
-    zeroSlackOrgStatus({
+    slackOrgStatus({
       orgId: auth.orgId,
       userId: auth.userId,
       orgRole: auth.orgRole,
@@ -628,7 +628,7 @@ const getSlackDownloadFileInner$ = computed(async (get) => {
   }
 
   const installation = await get(
-    zeroSlackOrgInstallation({ orgId: auth.orgId, userId: auth.userId }),
+    slackOrgInstallation({ orgId: auth.orgId, userId: auth.userId }),
   );
   if (!installation) {
     return jsonErrorResponse(
