+++ turbo/packages/db/scripts/agent-compose-consolidation-preflight-manifest.ts
@@ -203,6 +203,9 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/org-limited-free-bootstrap.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/org-limited-free-bootstrap.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/shared-thread.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
+    "turbo/apps/api/src/signals/services/slack-connect.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
+    "turbo/apps/api/src/signals/services/slack-connect.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
+    "turbo/apps/api/src/signals/services/slack-data.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/slack-message-context.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/slack-webhooks.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/teams-dispatch.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
@@ -223,9 +226,6 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/zero-chat-events.command.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/zero-runs-create.service.ts|@okouai/db/schema/agent-compose|agentComposeVersions:agentComposeVersions,agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/zero-runs-create.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
-    "turbo/apps/api/src/signals/services/zero-slack-connect.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
-    "turbo/apps/api/src/signals/services/zero-slack-connect.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
-    "turbo/apps/api/src/signals/services/zero-slack-data.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/zero-teams-connect.service.ts|@okouai/db/schema/agent-compose|agentComposeVersions:agentComposeVersions,agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/zero-teams-connect.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/zero-telegram-post.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
@@ -329,6 +329,8 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/run-mcp-connectors.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/shared-thread.service.ts|agentComposeId,agentComposes",
     "turbo/apps/api/src/signals/services/slack-chat-ingress.service.ts|agentComposeId",
+    "turbo/apps/api/src/signals/services/slack-connect.service.ts|agentComposes,zeroAgents",
+    "turbo/apps/api/src/signals/services/slack-data.service.ts|zeroAgents",
     "turbo/apps/api/src/signals/services/slack-message-context.service.ts|agentComposeId,zeroAgents",
     "turbo/apps/api/src/signals/services/slack-webhooks.service.ts|agentComposeId,zeroAgents",
     "turbo/apps/api/src/signals/services/storage-volume-publication.service.ts|headVersionId",
@@ -352,8 +354,6 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/zero-chat-events.command.ts|agentComposeId,zeroAgents",
     "turbo/apps/api/src/signals/services/zero-chat-title.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/zero-runs-create.service.ts|agentComposeId,agentComposeVersions,agentComposes,headVersionId,zeroAgents",
-    "turbo/apps/api/src/signals/services/zero-slack-connect.service.ts|agentComposes,zeroAgents",
-    "turbo/apps/api/src/signals/services/zero-slack-data.service.ts|zeroAgents",
     "turbo/apps/api/src/signals/services/zero-teams-connect.service.ts|agentComposeVersions,agentComposes,headVersionId,zeroAgents",
     "turbo/apps/api/src/signals/services/zero-telegram-post.service.ts|agentComposeId,agentComposes,zeroAgents",
     "turbo/apps/platform/src/views/zero-page/zero-telegram-settings-page.tsx|headVersionId",
